WT-17974 Add multi-node schema abort test infrastructure - iteration 2
